ow-pipeline-cdk/requirements.txt
aws-cdk.core>=1.132.0
aws-cdk.assertions>=1.132.0
aws-cdk.aws-codebuild>=1.132.0
aws-cdk.aws-iam>=1.132.0
aws-cdk.aws-s3>=1.132.0
